example/requirements.txt
typer~=0.4
ecs-logging
structlog~=21.5
sentineleof